update: structural chnages
