+++ src/components/InfiniteCarouselSection/InfiniteCarouselSection.jsx
@@ -1,0 +1,145 @@
+import React from "react";
+
+import "./InfiniteCarouselSection.css";
+
+const InfiniteCarouselSection = () => {
+  return (
+    <>
+      <div className="h4b_reg_text_wrapper">
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+
+        <p>Registrations are now Open !</p>
+      </div>
+    </>
+  );
+};
+
+export default InfiniteCarouselSection;
