+++ src/sections/PartnersAndSponsorsSection/PartnersAndSponsorsSection.jsx
@@ -5,50 +5,53 @@ import {motion} from "framer-motion";
 
 const PartnersAndSponsorsSection = () => {
   return (
-    <motion.div
-      initial={{opacity: 0}}
-      whileInView={{opacity: 1, transition: {duration: 1}}}
-      viewport={{once: false}}
-      className="hack4bengal__section6"
-      id="sponsers"
-    >
-      <div className="hack4bengal__section6-body">
-        <div className="hack4bengal__section6-title">Partners And Sponsors</div>
-        <hr />
-        <div className="hack4bengal__section6-content">
-          On behalf of the Hack4Bengal team, we would like to extend our
-          heartfelt gratitude to the partners and sponsors who will be
-          generously supporting us for the upcoming event. Their valuable
-          contributions will play an instrumental role in ensuring the smooth
-          running of the event, and we look forward to their support.
-        </div>
-        <div className="hack4bengal__section6-sponsers">
-          <div
-            className="hack4bengal__section6-sponser-title"
-            style={{marginBottom: "-40px"}}
-          >
-            In Collaboration With
+    <>
+      <motion.div
+        initial={{opacity: 0}}
+        whileInView={{opacity: 1, transition: {duration: 1}}}
+        viewport={{once: false}}
+        className="hack4bengal__section6"
+        id="sponsers"
+      >
+        <div className="hack4bengal__section6-body">
+          <div className="hack4bengal__section6-title">
+            Partners And Sponsors
           </div>
-          <div className="hack4bengal__section6-sponser-list platinum">
-            <div className="hack4bengal__section6-sponser-image">
-              {sponsers.platinum.map((sponser, index) => (
-                <a href={sponser.link}>
-                  <img key={index} src={sponser.img} />
-                  <div style={{textAlign: "center"}}>
-                    <p>
-                      {sponser.txt.split("\n").map((line, index) => (
-                        <React.Fragment key={index}>
-                          {line}
-                          <br />
-                        </React.Fragment>
-                      ))}
-                    </p>
-                  </div>
-                </a>
-              ))}
+          <hr />
+          <div className="hack4bengal__section6-content">
+            On behalf of the Hack4Bengal team, we would like to extend our
+            heartfelt gratitude to the partners and sponsors who will be
+            generously supporting us for the upcoming event. Their valuable
+            contributions will play an instrumental role in ensuring the smooth
+            running of the event, and we look forward to their support.
+          </div>
+          <div className="hack4bengal__section6-sponsers">
+            <div
+              className="hack4bengal__section6-sponser-title"
+              style={{marginBottom: "-40px"}}
+            >
+              In Collaboration With
             </div>
-          </div>
-          {/* <div className="hack4bengal__section6-sponser-title">
+            <div className="hack4bengal__section6-sponser-list platinum">
+              <div className="hack4bengal__section6-sponser-image">
+                {sponsers.platinum.map((sponser, index) => (
+                  <a href={sponser.link}>
+                    <img key={index} src={sponser.img} />
+                    <div style={{textAlign: "center"}}>
+                      <p>
+                        {sponser.txt.split("\n").map((line, index) => (
+                          <React.Fragment key={index}>
+                            {line}
+                            <br />
+                          </React.Fragment>
+                        ))}
+                      </p>
+                    </div>
+                  </a>
+                ))}
+              </div>
+            </div>
+            {/* <div className="hack4bengal__section6-sponser-title">
             Silver Sponsor
           </div>
           <div className="hack4bengal__section6-sponser-list platinum">
@@ -172,7 +175,7 @@ const PartnersAndSponsorsSection = () => {
               ))}
             </div>
           </div> */}
-          {/*   {sponsers.platinum.map((sponser, index) => (
+            {/*   {sponsers.platinum.map((sponser, index) => (
               <div className="hack4bengal__section6-sponser-image">
                 <a href={sponser.link}>
                   <img key={index} src={sponser.img} />
@@ -183,7 +186,7 @@ const PartnersAndSponsorsSection = () => {
               </div>
             ))} */}
 
-          {/* <div className="hack4bengal__section6-sponser-title">
+            {/* <div className="hack4bengal__section6-sponser-title">
             Incubation Partner
           </div>
           <div className="hack4bengal__section6-sponser-list gold">
@@ -198,77 +201,6 @@ const PartnersAndSponsorsSection = () => {
             
           </div> */}
 
-          <div
-            className="two-group-class"
-            style={{
-              display: "flex",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "center",
-              width: "550px",
-            }}
-          >
-            <div className="hack4bengal__section6-sponser-title sponser-two-dual">
-              Gold Sponsor
-            </div>
-            <div className="hack4bengal__section6-sponser-list platinum">
-              <div className="hack4bengal__section6-sponser-image">
-                {sponsers.gold.map((sponser, index) => (
-                  <a href={sponser.link}>
-                    <img
-                      style={{borderRadius: "20px"}}
-                      key={index}
-                      src={sponser.img}
-                    />
-                    <div style={{textAlign: "center"}}>
-                      <p style={{paddingTop: "2px"}}>
-                        {sponser.txt.split("\n").map((line, index) => (
-                          <React.Fragment key={index}>{line}</React.Fragment>
-                        ))}
-                      </p>
-                    </div>
-                  </a>
-                ))}
-              </div>
-            </div>
-          </div>
-
-          <div
-            className="two-group-class"
-            style={{
-              display: "flex",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "center",
-              width: "550px",
-            }}
-          >
-            <div className="hack4bengal__section6-sponser-title sponser-two-dual">
-              Silver Sponsor
-            </div>
-            <div className="hack4bengal__section6-sponser-list platinum">
-              <div className="hack4bengal__section6-sponser-image">
-                {sponsers.silver.map((sponser, index) => (
-                  <a href={sponser.link}>
-                    <img
-                      style={{borderRadius: "20px"}}
-                      key={index}
-                      src={sponser.img}
-                    />
-                    <div style={{textAlign: "center"}}>
-                      <p style={{paddingTop: "2px"}}>
-                        {sponser.txt.split("\n").map((line, index) => (
-                          <React.Fragment key={index}>{line}</React.Fragment>
-                        ))}
-                      </p>
-                    </div>
-                  </a>
-                ))}
-              </div>
-            </div>
-          </div>
-
-          <div className="partner-two-group">
             <div
               className="two-group-class"
               style={{
@@ -280,42 +212,11 @@ const PartnersAndSponsorsSection = () => {
               }}
             >
               <div className="hack4bengal__section6-sponser-title sponser-two-dual">
-                Bronze Sponsor
+                Gold Sponsor
               </div>
               <div className="hack4bengal__section6-sponser-list platinum">
                 <div className="hack4bengal__section6-sponser-image">
-                  {sponsers.bronze.map((sponser, index) => (
-                    <a href={sponser.link}>
-                      <img key={index} src={sponser.img} />
-                      <div style={{textAlign: "center"}}>
-                        <p style={{paddingTop: "2px"}}>
-                          {sponser.txt.split("\n").map((line, index) => (
-                            <React.Fragment key={index}>{line}</React.Fragment>
-                          ))}
-                        </p>
-                      </div>
-                    </a>
-                  ))}
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="partner-two-group">
-            <div
-              className="two-group-class"
-              style={{
-                display: "flex",
-                flexDirection: "column",
-                justifyContent: "center",
-                alignItems: "center",
-              }}
-            >
-              <div className="hack4bengal__section6-sponser-title sponser-two-dual">
-                Digital Media Partner
-              </div>
-              <div className="hack4bengal__section6-sponser-list platinum">
-                <div className="hack4bengal__section6-sponser-image">
-                  {sponsers.media.map((sponser, index) => (
+                  {sponsers.gold.map((sponser, index) => (
                     <a href={sponser.link}>
                       <img
                         style={{borderRadius: "20px"}}
@@ -334,42 +235,7 @@ const PartnersAndSponsorsSection = () => {
                 </div>
               </div>
             </div>
-            <div
-              className="two-group-class"
-              style={{
-                display: "flex",
-                flexDirection: "column",
-                justifyContent: "center",
-                alignItems: "center",
-              }}
-            >
-              <div className="hack4bengal__section6-sponser-title sponser-two-dual">
-                Domain Partner
-              </div>
-              <div className="hack4bengal__section6-sponser-list platinum">
-                <div className="hack4bengal__section6-sponser-image">
-                  {sponsers.domain.map((sponser, index) => (
-                    <a href={sponser.link}>
-                      <img
-                        style={{borderRadius: "20px", width: "158px"}}
-                        key={index}
-                        src={sponser.img}
-                      />
-                      <div style={{textAlign: "center"}}>
-                        <p style={{paddingTop: "2px"}}>
-                          {sponser.txt.split("\n").map((line, index) => (
-                            <React.Fragment key={index}>{line}</React.Fragment>
-                          ))}
-                        </p>
-                      </div>
-                    </a>
-                  ))}
-                </div>
-              </div>
-            </div>
-          </div>
 
-          <div className="partner-two-group">
             <div
               className="two-group-class"
               style={{
@@ -377,22 +243,18 @@ const PartnersAndSponsorsSection = () => {
                 flexDirection: "column",
                 justifyContent: "center",
                 alignItems: "center",
+                width: "550px",
               }}
             >
               <div className="hack4bengal__section6-sponser-title sponser-two-dual">
-                Energy Partner
+                Silver Sponsor
               </div>
               <div className="hack4bengal__section6-sponser-list platinum">
                 <div className="hack4bengal__section6-sponser-image">
-                  {sponsers.energy.map((sponser, index) => (
+                  {sponsers.silver.map((sponser, index) => (
                     <a href={sponser.link}>
                       <img
-                        style={{
-                          borderRadius: "20px",
-                          width: "240px",
-                          height: "100px",
-                          objectFit: "contain",
-                        }}
+                        style={{borderRadius: "20px"}}
                         key={index}
                         src={sponser.img}
                       />
@@ -408,42 +270,193 @@ const PartnersAndSponsorsSection = () => {
                 </div>
               </div>
             </div>
-            <div
-              className="two-group-class"
-              style={{
-                display: "flex",
-                flexDirection: "column",
-                justifyContent: "center",
-                alignItems: "center",
-              }}
-            >
-              <div className="hack4bengal__section6-sponser-title sponser-two-dual">
-                Social Media Partner
-              </div>
-              <div className="hack4bengal__section6-sponser-list platinum">
-                <div className="hack4bengal__section6-sponser-image">
-                  {sponsers.social.map((sponser, index) => (
-                    <a href={sponser.link}>
-                      <img
-                        style={{borderRadius: "20px", width: "100px"}}
-                        key={index}
-                        src={sponser.img}
-                      />
-                      <div style={{textAlign: "center"}}>
-                        <p style={{paddingTop: "2px"}}>
-                          {sponser.txt.split("\n").map((line, index) => (
-                            <React.Fragment key={index}>{line}</React.Fragment>
-                          ))}
-                        </p>
-                      </div>
-                    </a>
-                  ))}
-                </div>
-              </div>
-            </div>
-          </div>
 
-          {/*   <div
+            <div className="partner-two-group">
+              <div
+                className="two-group-class"
+                style={{
+                  display: "flex",
+                  flexDirection: "column",
+                  justifyContent: "center",
+                  alignItems: "center",
+                  width: "550px",
+                }}
+              >
+                <div className="hack4bengal__section6-sponser-title sponser-two-dual">
+                  Bronze Sponsor
+                </div>
+                <div className="hack4bengal__section6-sponser-list platinum">
+                  <div className="hack4bengal__section6-sponser-image">
+                    {sponsers.bronze.map((sponser, index) => (
+                      <a href={sponser.link}>
+                        <img key={index} src={sponser.img} />
+                        <div style={{textAlign: "center"}}>
+                          <p style={{paddingTop: "2px"}}>
+                            {sponser.txt.split("\n").map((line, index) => (
+                              <React.Fragment key={index}>
+                                {line}
+                              </React.Fragment>
+                            ))}
+                          </p>
+                        </div>
+                      </a>
+                    ))}
+                  </div>
+                </div>
+              </div>
+            </div>
+            <div className="partner-two-group">
+              <div
+                className="two-group-class"
+                style={{
+                  display: "flex",
+                  flexDirection: "column",
+                  justifyContent: "center",
+                  alignItems: "center",
+                }}
+              >
+                <div className="hack4bengal__section6-sponser-title sponser-two-dual">
+                  Digital Media Partner
+                </div>
+                <div className="hack4bengal__section6-sponser-list platinum">
+                  <div className="hack4bengal__section6-sponser-image">
+                    {sponsers.media.map((sponser, index) => (
+                      <a href={sponser.link}>
+                        <img
+                          style={{borderRadius: "20px"}}
+                          key={index}
+                          src={sponser.img}
+                        />
+                        <div style={{textAlign: "center"}}>
+                          <p style={{paddingTop: "2px"}}>
+                            {sponser.txt.split("\n").map((line, index) => (
+                              <React.Fragment key={index}>
+                                {line}
+                              </React.Fragment>
+                            ))}
+                          </p>
+                        </div>
+                      </a>
+                    ))}
+                  </div>
+                </div>
+              </div>
+              <div
+                className="two-group-class"
+                style={{
+                  display: "flex",
+                  flexDirection: "column",
+                  justifyContent: "center",
+                  alignItems: "center",
+                }}
+              >
+                <div className="hack4bengal__section6-sponser-title sponser-two-dual">
+                  Domain Partner
+                </div>
+                <div className="hack4bengal__section6-sponser-list platinum">
+                  <div className="hack4bengal__section6-sponser-image">
+                    {sponsers.domain.map((sponser, index) => (
+                      <a href={sponser.link}>
+                        <img
+                          style={{borderRadius: "20px", width: "158px"}}
+                          key={index}
+                          src={sponser.img}
+                        />
+                        <div style={{textAlign: "center"}}>
+                          <p style={{paddingTop: "2px"}}>
+                            {sponser.txt.split("\n").map((line, index) => (
+                              <React.Fragment key={index}>
+                                {line}
+                              </React.Fragment>
+                            ))}
+                          </p>
+                        </div>
+                      </a>
+                    ))}
+                  </div>
+                </div>
+              </div>
+            </div>
+
+            <div className="partner-two-group">
+              <div
+                className="two-group-class"
+                style={{
+                  display: "flex",
+                  flexDirection: "column",
+                  justifyContent: "center",
+                  alignItems: "center",
+                }}
+              >
+                <div className="hack4bengal__section6-sponser-title sponser-two-dual">
+                  Energy Partner
+                </div>
+                <div className="hack4bengal__section6-sponser-list platinum">
+                  <div className="hack4bengal__section6-sponser-image">
+                    {sponsers.energy.map((sponser, index) => (
+                      <a href={sponser.link}>
+                        <img
+                          style={{
+                            borderRadius: "20px",
+                            width: "240px",
+                            height: "100px",
+                            objectFit: "contain",
+                          }}
+                          key={index}
+                          src={sponser.img}
+                        />
+                        <div style={{textAlign: "center"}}>
+                          <p style={{paddingTop: "2px"}}>
+                            {sponser.txt.split("\n").map((line, index) => (
+                              <React.Fragment key={index}>
+                                {line}
+                              </React.Fragment>
+                            ))}
+                          </p>
+                        </div>
+                      </a>
+                    ))}
+                  </div>
+                </div>
+              </div>
+              <div
+                className="two-group-class"
+                style={{
+                  display: "flex",
+                  flexDirection: "column",
+                  justifyContent: "center",
+                  alignItems: "center",
+                }}
+              >
+                <div className="hack4bengal__section6-sponser-title sponser-two-dual">
+                  Social Media Partner
+                </div>
+                <div className="hack4bengal__section6-sponser-list platinum">
+                  <div className="hack4bengal__section6-sponser-image">
+                    {sponsers.social.map((sponser, index) => (
+                      <a href={sponser.link}>
+                        <img
+                          style={{borderRadius: "20px", width: "100px"}}
+                          key={index}
+                          src={sponser.img}
+                        />
+                        <div style={{textAlign: "center"}}>
+                          <p style={{paddingTop: "2px"}}>
+                            {sponser.txt.split("\n").map((line, index) => (
+                              <React.Fragment key={index}>
+                                {line}
+                              </React.Fragment>
+                            ))}
+                          </p>
+                        </div>
+                      </a>
+                    ))}
+                  </div>
+                </div>
+              </div>
+            </div>
+
+            {/*   <div
             className="hack4bengal__section6-sponser-title"
             style={{ marginBottom: "-40px" }}
           >
@@ -509,47 +522,47 @@ const PartnersAndSponsorsSection = () => {
             </div>
           </div> */}
 
-          <div
-            className="two-group-class"
-            style={{
-              display: "flex",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "center",
-              width: "100%",
-            }}
-          >
-            <div className="hack4bengal__section6-sponser-title sponser-two-dual">
-              In-Kind Sponsor
-            </div>
-            <div className="hack4bengal__section6-sponser-list platinum">
-              <div className="hack4bengal__section6-sponser-image">
-                {sponsers.inkind.map((sponser, index) => (
-                  <a href={sponser.link}>
-                    <img
-                      style={{
-                        borderRadius: "20px",
-                        width: "200px",
-                        height: "100px",
-                        objectFit: "contain",
-                      }}
-                      key={index}
-                      src={sponser.img}
-                    />
-                    <div style={{textAlign: "center"}}>
-                      <p style={{paddingTop: "2px"}}>
-                        {sponser.txt.split("\n").map((line, index) => (
-                          <React.Fragment key={index}>{line}</React.Fragment>
-                        ))}
-                      </p>
-                    </div>
-                  </a>
-                ))}
+            <div
+              className="two-group-class"
+              style={{
+                display: "flex",
+                flexDirection: "column",
+                justifyContent: "center",
+                alignItems: "center",
+                width: "100%",
+              }}
+            >
+              <div className="hack4bengal__section6-sponser-title sponser-two-dual">
+                In-Kind Sponsor
+              </div>
+              <div className="hack4bengal__section6-sponser-list platinum">
+                <div className="hack4bengal__section6-sponser-image">
+                  {sponsers.inkind.map((sponser, index) => (
+                    <a href={sponser.link}>
+                      <img
+                        style={{
+                          borderRadius: "20px",
+                          width: "200px",
+                          height: "100px",
+                          objectFit: "contain",
+                        }}
+                        key={index}
+                        src={sponser.img}
+                      />
+                      <div style={{textAlign: "center"}}>
+                        <p style={{paddingTop: "2px"}}>
+                          {sponser.txt.split("\n").map((line, index) => (
+                            <React.Fragment key={index}>{line}</React.Fragment>
+                          ))}
+                        </p>
+                      </div>
+                    </a>
+                  ))}
+                </div>
               </div>
             </div>
-          </div>
 
-          {/* 
+            {/* 
           <div className="hack4bengal__section6-sponser-list platinum">
             <div className="hack4bengal__section6-sponser-image">
               {sponsers.domain.map((sponser, index) => (
@@ -568,24 +581,25 @@ const PartnersAndSponsorsSection = () => {
               ))}
             </div>
           </div> */}
-        </div>
-        <h2>More Sponsors To be Declared Soon</h2>
+          </div>
+          <h2>More Sponsors To be Declared Soon</h2>
 
-        <div
-          className="hack4bengal__section6-content"
-          style={{marginTop: "50px", fontWeight: "bold"}}
-        >
-          Want to sponsor us? <br />
-          <a
-            href="https://forms.gle/Y77QCcKmMAMpPAmQ9"
-            className="link-us-sponsor"
-            style={{color: "#E13B3C"}}
+          <div
+            className="hack4bengal__section6-content"
+            style={{marginTop: "50px", fontWeight: "bold"}}
           >
-            Contact us
-          </a>
+            Want to sponsor us? <br />
+            <a
+              href="https://forms.gle/Y77QCcKmMAMpPAmQ9"
+              className="link-us-sponsor"
+              style={{color: "#E13B3C"}}
+            >
+              Contact us
+            </a>
+          </div>
         </div>
-      </div>
-    </motion.div>
+      </motion.div>
+    </>
   );
 };
