+++ src/sections/PrizesSection/PrizesSection.jsx
