+++ src/sections/WelcomeSection/WelcomeSection.jsx
@@ -1,184 +1,211 @@
-import React from "react";
-import "./Footer.css";
-import footer from "../../assets/img/footer.gif";
-import { Link } from "react-router-dom";
+import React, {useEffect} from "react";
+import "./WelcomeSection.css";
+import h4blogo from "../../assets/img/square-grey-logo.png";
+import discord from "../../assets/img/discord-mark-blue.png";
+import newbanner from "../../assets/updates/newbanner.png";
+import newbanner2 from "../../assets/updates/newbanner3.png";
+import {motion, transform} from "framer-motion";
+import {staggerChildren, fadeIn} from "../../motionUtils";
+import "../../../node_modules/font-awesome/css/font-awesome.min.css";
 
-const Footer = () => {
+const WelcomeSection = () => {
+  React.useEffect(() => {
+    const script = document.createElement("script");
+    script.src = "https://apply.devfolio.co/v2/sdk.js";
+    script.async = true;
+    script.defer = true;
+    document.body.appendChild(script);
+    return () => {
+      document.body.removeChild(script);
+    };
+  }, []);
+
   return (
-    <div className="hack4bengal__footer" id="contactus">
-      <div className="hack4bengal__footer-container1">
-        <div className="hack4bengal__footer-logos">
-          <img
-            src={footer}
-            className="hack4bengal__footer-logo-tcp"
-            alt="Hack4Bengal"
-          />
-        </div>
-        <div className="hack4bengal__footer-social-container">
-          <div className="hack4bengal__footer-social-links-container">
+    <>
+      <motion.div
+        initial="hidden"
+        whileInView="show"
+        variants={staggerChildren}
+        viewport={{once: true, amount: 0.25}}
+        className="hack4bengal__section1"
+        id="home"
+      >
+        <motion.div
+          variants={fadeIn("down", "tween", 0.5, 1)}
+          className="hack4bengal__section1-body"
+        >
+          <div className="hack4bengal__section1-menu">
+            <div className="hack4bengal__section1-heading">Welcome To</div>
+
+            <div className="hack4bengal_section1_header">
+              <img src={newbanner2} alt="hack4bengal"></img>
+            </div>
+
+            <div className="hack4bengal__section1-para">
+              <p className="span1"> In Collaboration with</p>{" "}
+              <p className="span2">
+                Cyber Security Center of Excellence <br /> Department of
+                Information Technology & Electronics <br /> Government of West
+                Bengal
+              </p>
+            </div>
+          </div>
+          <div
+            className="hack4bengal__section1-button-container"
+            style={{display: "flex", placeContent: "center"}}
+          >
+            <div className="hack4bengal__section1-discord" id="discord">
+              <a
+                className="hack4bengal__section1-discord-link landingpagebtn"
+                href="https://discord.io/Hack4Bengal"
+                target="_blank"
+                rel="noreferrer"
+              >
+                <img src={discord} alt="discord" />
+                {"  "}
+                <p>Join Discord</p>
+              </a>
+            </div>
+            {/* <div className="hack4bengal__section1-register">
+            <a
+              className="hack4bengal__section1-register-link"
+              href="https://forms.gle/2Ehe6UyD4BLisVRr5"
+              target="_blank"
+              rel="noreferrer"
+            >
+              <span>{"</> "}</span>Coming Soon
+            </a>
+          </div> */}
+            {/* 
+          <div className="hack4bengal__section1-register">
+            <a className="hack4bengal__section1-register-link" href="#">
+              <span>{"</> "}</span>Coming Soon
+            </a>
+          </div> */}
+
+            <div className="hack4bengal__section1-register">
+              <a
+                className="hack4bengal__section1-register-link landingpagebtn"
+                href="https://lu.ma/hack4bengal"
+                target="_blank"
+                rel="noreferrer"
+              >
+                <span>{"</> "}</span>
+                <p>Register Now</p>
+              </a>
+            </div>
+          </div>
+        </motion.div>
+        <motion.ul
+          variants={fadeIn("up", "tween", 1, 1)}
+          className="hack4bengal__section1-side-links"
+        >
+          <li>
             <a
               target="_blank"
               rel="noreferrer"
               href="https://www.facebook.com/hack4bengal/"
             >
               <svg
-                fill="#999"
-                width="32"
-                height="32"
+                fill="#eee"
+                width="28"
+                height="28"
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 512 512"
               >
                 <path d="M504 256C504 119 393 8 256 8S8 119 8 256c0 123.78 90.69 226.38 209.25 245V327.69h-63V256h63v-54.64c0-62.15 37-96.48 93.67-96.48 27.14 0 55.52 4.84 55.52 4.84v61h-31.28c-30.8 0-40.41 19.12-40.41 38.73V256h68.78l-11 71.69h-57.78V501C413.31 482.38 504 379.78 504 256z" />
               </svg>
             </a>
+          </li>
+          <li>
             <a
               target="_blank"
               rel="noreferrer"
               href="https://www.instagram.com/hack4bengal/"
             >
               <svg
-                fill="#999"
-                width="32"
-                height="32"
+                fill="#eee"
+                width="28"
+                height="28"
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 448 512"
               >
                 <path d="M224.1 141c-63.6 0-114.9 51.3-114.9 114.9s51.3 114.9 114.9 114.9S339 319.5 339 255.9 287.7 141 224.1 141zm0 189.6c-41.1 0-74.7-33.5-74.7-74.7s33.5-74.7 74.7-74.7 74.7 33.5 74.7 74.7-33.6 74.7-74.7 74.7zm146.4-194.3c0 14.9-12 26.8-26.8 26.8-14.9 0-26.8-12-26.8-26.8s12-26.8 26.8-26.8 26.8 12 26.8 26.8zm76.1 27.2c-1.7-35.9-9.9-67.7-36.2-93.9-26.2-26.2-58-34.4-93.9-36.2-37-2.1-147.9-2.1-184.9 0-35.8 1.7-67.6 9.9-93.9 36.1s-34.4 58-36.2 93.9c-2.1 37-2.1 147.9 0 184.9 1.7 35.9 9.9 67.7 36.2 93.9s58 34.4 93.9 36.2c37 2.1 147.9 2.1 184.9 0 35.9-1.7 67.7-9.9 93.9-36.2 26.2-26.2 34.4-58 36.2-93.9 2.1-37 2.1-147.8 0-184.8zM398.8 388c-7.8 19.6-22.9 34.7-42.6 42.6-29.5 11.7-99.5 9-132.1 9s-102.7 2.6-132.1-9c-19.6-7.8-34.7-22.9-42.6-42.6-11.7-29.5-9-99.5-9-132.1s-2.6-102.7 9-132.1c7.8-19.6 22.9-34.7 42.6-42.6 29.5-11.7 99.5-9 132.1-9s102.7-2.6 132.1 9c19.6 7.8 34.7 22.9 42.6 42.6 11.7 29.5 9 99.5 9 132.1s2.7 102.7-9 132.1z" />
               </svg>
             </a>
+          </li>
+          <li>
             <a
               target="_blank"
               rel="noreferrer"
               href="https://github.com/hack4bengal"
             >
               <svg
-                fill="#999"
-                width="32"
-                height="32"
+                fill="#eee"
+                width="28"
+                height="28"
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 496 512"
               >
                 <path d="M165.9 397.4c0 2-2.3 3.6-5.2 3.6-3.3.3-5.6-1.3-5.6-3.6 0-2 2.3-3.6 5.2-3.6 3-.3 5.6 1.3 5.6 3.6zm-31.1-4.5c-.7 2 1.3 4.3 4.3 4.9 2.6 1 5.6 0 6.2-2s-1.3-4.3-4.3-5.2c-2.6-.7-5.5.3-6.2 2.3zm44.2-1.7c-2.9.7-4.9 2.6-4.6 4.9.3 2 2.9 3.3 5.9 2.6 2.9-.7 4.9-2.6 4.6-4.6-.3-1.9-3-3.2-5.9-2.9zM244.8 8C106.1 8 0 113.3 0 252c0 110.9 69.8 205.8 169.5 239.2 12.8 2.3 17.3-5.6 17.3-12.1 0-6.2-.3-40.4-.3-61.4 0 0-70 15-84.7-29.8 0 0-11.4-29.1-27.8-36.6 0 0-22.9-15.7 1.6-15.4 0 0 24.9 2 38.6 25.8 21.9 38.6 58.6 27.5 72.9 20.9 2.3-16 8.8-27.1 16-33.7-55.9-6.2-112.3-14.3-112.3-110.5 0-27.5 7.6-41.3 23.6-58.9-2.6-6.5-11.1-33.3 2.6-67.9 20.9-6.5 69 27 69 27 20-5.6 41.5-8.5 62.8-8.5s42.8 2.9 62.8 8.5c0 0 48.1-33.6 69-27 13.7 34.7 5.2 61.4 2.6 67.9 16 17.7 25.8 31.5 25.8 58.9 0 96.5-58.9 104.2-114.8 110.5 9.2 7.9 17 22.9 17 46.4 0 33.7-.3 75.4-.3 83.6 0 6.5 4.6 14.4 17.3 12.1C428.2 457.8 496 362.9 496 252 496 113.3 383.5 8 244.8 8zM97.2 352.9c-1.3 1-1 3.3.7 5.2 1.6 1.6 3.9 2.3 5.2 1 1.3-1 1-3.3-.7-5.2-1.6-1.6-3.9-2.3-5.2-1zm-10.8-8.1c-.7 1.3.3 2.9 2.3 3.9 1.6 1 3.6.7 4.3-.7.7-1.3-.3-2.9-2.3-3.9-2-.6-3.6-.3-4.3.7zm32.4 35.6c-1.6 1.3-1 4.3 1.3 6.2 2.3 2.3 5.2 2.6 6.5 1 1.3-1.3.7-4.3-1.3-6.2-2.2-2.3-5.2-2.6-6.5-1zm-11.4-14.7c-1.6 1-1.6 3.6 0 5.9 1.6 2.3 4.3 3.3 5.6 2.3 1.6-1.3 1.6-3.9 0-6.2-1.4-2.3-4-3.3-5.6-2z" />
               </svg>
             </a>
+          </li>
+          <li>
             <a
               target="_blank"
               rel="noreferrer"
               href="https://twitter.com/hack4bengal"
             >
               <svg
-                fill="#999"
-                width="32"
-                height="32"
+                fill="#eee"
+                width="28"
+                height="28"
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 448 512"
               >
                 <path d="M400 32H48C21.5 32 0 53.5 0 80v352c0 26.5 21.5 48 48 48h352c26.5 0 48-21.5 48-48V80c0-26.5-21.5-48-48-48zm-48.9 158.8c.2 2.8.2 5.7.2 8.5 0 86.7-66 186.6-186.6 186.6-37.2 0-71.7-10.8-100.7-29.4 5.3.6 10.4.8 15.8.8 30.7 0 58.9-10.4 81.4-28-28.8-.6-53-19.5-61.3-45.5 10.1 1.5 19.2 1.5 29.6-1.2-30-6.1-52.5-32.5-52.5-64.4v-.8c8.7 4.9 18.9 7.9 29.6 8.3a65.447 65.447 0 0 1-29.2-54.6c0-12.2 3.2-23.4 8.9-33.1 32.3 39.8 80.8 65.8 135.2 68.6-9.3-44.5 24-80.6 64-80.6 18.9 0 35.9 7.9 47.9 20.7 14.8-2.8 29-8.3 41.6-15.8-4.9 15.2-15.2 28-28.8 36.1 13.2-1.4 26-5.1 37.8-10.2-8.9 13.1-20.1 24.7-32.9 34z" />
               </svg>
             </a>
+          </li>
+          <li>
             <a
-              target="linkedin"
+              target="_blank"
+              rel="noreferrer"
               href="https://www.linkedin.com/company/hack4bengal/"
             >
               <svg
-                fill="#999"
-                width="32"
-                height="32"
+                fill="#eee"
+                width="28"
+                height="28"
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 448 512"
               >
                 <path d="M416 32H31.9C14.3 32 0 46.5 0 64.3v383.4C0 465.5 14.3 480 31.9 480H416c17.6 0 32-14.5 32-32.3V64.3c0-17.8-14.4-32.3-32-32.3zM135.4 416H69V202.2h66.5V416zm-33.2-243c-21.3 0-38.5-17.3-38.5-38.5S80.9 96 102.2 96c21.2 0 38.5 17.3 38.5 38.5 0 21.3-17.2 38.5-38.5 38.5zm282.1 243h-66.4V312c0-24.8-.5-56.7-34.5-56.7-34.6 0-39.9 27-39.9 54.9V416h-66.4V202.2h63.7v29.2h.9c8.9-16.8 30.6-34.5 62.9-34.5 67.2 0 79.7 44.3 79.7 101.9V416z" />
               </svg>
             </a>
-          </div>
-          {window.location.pathname !== "/coc" && (
-            <p className="hack4bengal__footer-social-text" style={{ marginTop: '20px' }}>
-              <Link to="/coc">Code of Conduct</Link>
-            </p>
-          )}
-
-          <p className="hack4bengal__footer-social-text">
-            Email us for more details,
-          </p>
-          <p className="hack4bengal__footer-social-text">
-            <a href="mailto:hack4bengal@gmail.com" style={{ color: "#e13b3c" }}>
-              hack4bengal@gmail.com
-            </a>
-          </p>
-        </div>
-      </div>
-
-      <div className="hack4bengal__footer-container3">
-        <div className="hack4bengal__footer-social-team">
-          Developed By <Link to="/team">Hack4Bengal</Link>
-        </div>
-        <div
-          className="hack4bengal__footer-back-to-top"
-          onClick={() => {
-            window.scrollTo(0, 0);
-          }}
+          </li>
+        </motion.ul>
+        <motion.div
+          variants={fadeIn("up", "tween", 1.2, 1)}
+          className="hack4bengal__section1-scrollDown"
         >
-          <p>
-            Back to Top
-            <svg
-              fill="#ff1717"
-              height="20"
-              width="20"
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 320 512"
-              style={{ marginLeft: "6px", transform: "translateX(2px)" }}
-            >
-              <path d="M214.6 41.4c-12.5-12.5-32.8-12.5-45.3 0l-160 160c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0L160 141.2V448c0 17.7 14.3 32 32 32s32-14.3 32-32V141.2L329.4 246.6c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3l-160-160z" />
-            </svg>
-          </p>
-        </div>
-
-        <div className="hack4bengal__footer-social-map">
-          <a
-            href="https://s1.hack4bengal.tech/"
-            target="_blank"
-            rel="noreferrer"
-          >
-            <svg
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="#ff1717"
-              xmlns="http://www.w3.org/2000/svg"
-              style={{ marginRight: "6px", marginBottom: "-5px" }}
-            >
-              {" "}
-              <path
-                fill-rule="evenodd"
-                clip-rule="evenodd"
-                d="M14 7C13.4477 7 13 7.44772 13 8V16C13 16.5523 13.4477 17 14 17H18C18.5523 17 19 16.5523 19 16V8C19 7.44772 18.5523 7 18 7H14ZM17 9H15V15H17V9Z"
-                fill="red"
-              ></path>{" "}
-              <path
-                d="M6 7C5.44772 7 5 7.44772 5 8C5 8.55228 5.44772 9 6 9H10C10.5523 9 11 8.55228 11 8C11 7.44772 10.5523 7 10 7H6Z"
-                fill="red"
-              ></path>{" "}
-              <path
-                d="M6 11C5.44772 11 5 11.4477 5 12C5 12.5523 5.44772 13 6 13H10C10.5523 13 11 12.5523 11 12C11 11.4477 10.5523 11 10 11H6Z"
-                fill="red"
-              ></path>{" "}
-              <path
-                d="M5 16C5 15.4477 5.44772 15 6 15H10C10.5523 15 11 15.4477 11 16C11 16.5523 10.5523 17 10 17H6C5.44772 17 5 16.5523 5 16Z"
-                fill="red"
-              ></path>{" "}
-              <path
-                fill-rule="evenodd"
-                clip-rule="evenodd"
-                d="M4 3C2.34315 3 1 4.34315 1 6V18C1 19.6569 2.34315 21 4 21H20C21.6569 21 23 19.6569 23 18V6C23 4.34315 21.6569 3 20 3H4ZM20 5H4C3.44772 5 3 5.44772 3 6V18C3 18.5523 3.44772 19 4 19H20C20.5523 19 21 18.5523 21 18V6C21 5.44772 20.5523 5 20 5Z"
-                fill="red"
-              ></path>{" "}
-            </svg>
-            Visit Season 1
+          <a href="#aboutus">
+            <b>Scroll Down</b>
           </a>
-        </div>
-      </div>
-    </div>
+          <svg
+            fill="#ff1717"
+            width="19"
+            height="19"
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox="0 0 384 512"
+          >
+            <path d="M224 402.7V32c0-17.7-14.3-32-32-32s-32 14.3-32 32V402.7L86.6 329.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l128 128c12.5 12.5 32.8 12.5 45.3 0l128-128c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L224 402.7z" />
+          </svg>
+        </motion.div>
+      </motion.div>
+    </>
   );
 };
 
-export default Footer;
+export default WelcomeSection;
